+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/util/TableUtilsSpec.groovy
@@ -50,11 +50,6 @@ class TableUtilsSpec extends  Specification {
         metric3 = "m3"
     }
 
-    def setup() {
-        query.getDataSource() >> new TableDataSource(new PhysicalTable("table_name", DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC), [:]))
-    }
-
-
     @Unroll
     def "With #requestDimensions, #filterDimensions, #metricFilterDimensions dimension names are: #expected  "() {
         setup:
@@ -65,7 +60,7 @@ class TableUtilsSpec extends  Specification {
         expected = expected as Set
 
         expect:
-        TableUtils.getColumnNames(request, query, new PhysicalTable("", DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC), [:])) == expected
+        TableUtils.getColumnNames(request, query) == expected
 
         where:
         requestDimensions | filterDimensions | metricFilterDimensions | expected
@@ -87,19 +82,6 @@ class TableUtilsSpec extends  Specification {
         query.dependentFieldNames >> ([metric1, metric2, metric3] as Set)
 
         expect:
-        TableUtils.getColumnNames(request, query, new PhysicalTable("", DefaultTimeGrain.DAY.buildZonedTimeGrain(DateTimeZone.UTC), [:])) == [d1Name, metric1, metric2, metric3] as Set
-    }
-
-    def "logicalName to physicalName mapping not required for metrics" () {
-        setup:
-        request.dimensions >> []
-        request.filterDimensions >> []
-        query.metricDimensions >> []
-        query.dependentFieldNames >> ([metric1] as Set)
-        PhysicalTable physicalTable = Mock(PhysicalTable)
-        0 * physicalTable.getPhysicalColumnName(_)
-
-        expect:
-        TableUtils.getColumnNames(request, query, physicalTable) == [metric1] as Set
+        TableUtils.getColumnNames(request, query) == [d1Name, metric1, metric2, metric3] as Set
     }
 }